@@ -1,37 +1,38 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 26
+    compileSdkVersion 27
     defaultConfig {
         applicationId "com.noclip.marcinmalysz.pixfight"
         minSdkVersion 21
-        targetSdkVersion 26
-        versionCode 1
-        versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        targetSdkVersion 27
+        versionCode 2
+        versionName "1.1"
         externalNativeBuild {
             cmake {
                 cppFlags "-std=c++14 -frtti -fexceptions -Wno-narrowing"
                 arguments '-DANDROID_STL=c++_static'
             }
         }
+        resConfigs "en"
     }
     buildTypes {
 
         debug {
 
             ndk {
-                abiFilters 'x86', 'armeabi', 'armeabi-v7a', 'arm64-v8a'
+                abiFilters 'x86', 'armeabi-v7a', 'arm64-v8a'
             }
         }
 
         release {
 
             ndk {
-                abiFilters 'x86', 'armeabi', 'armeabi-v7a', 'arm64-v8a'
+                abiFilters 'x86', 'armeabi-v7a', 'arm64-v8a'
             }
 
-            minifyEnabled false
+            minifyEnabled true
+            shrinkResources true
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
@@ -40,15 +41,17 @@ android {
             path "CMakeLists.txt"
         }
     }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
 }
 
 dependencies {
-    compile files('../FMOD/lib/fmod.jar')
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:26.1.0'
-    implementation 'com.android.support.constraint:constraint-layout:1.0.2'
-    implementation 'com.android.support:support-v4:26.1.0'
-    testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.1'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.1'
+    implementation files('../FMOD/lib/fmod.jar')
+    implementation ('com.android.support:appcompat-v7:27.1.1') {
+        exclude module: 'support-v4'
+    }
+    implementation 'com.android.support.constraint:constraint-layout:1.1.0'
+    implementation 'com.android.support:support-fragment:27.1.1'
 }
